fix(firmware): resolve 16-bit task_id truncation and MIPC timeout deadlocks

+++ firmware/Core/Inc/scheduler.h
@@ -29,7 +29,7 @@ typedef struct {
     uint8_t minute;
     uint8_t second;
     TaskAction_t action;
-    uint16_t task_id;
+    uint32_t task_id;
     char objective[64];
 } ScheduledTask_t;
